Add date and text formatter utils
