+++ app/components/ConversationSidebar.tsx
@@ -2,6 +2,9 @@
 import { useState } from "react";
 import { Link, useNavigate } from "@remix-run/react";
 
+// Utils
+import { formatRelativeDate, truncateText } from "~/utils/format";
+
 // Types
 import type { Conversation } from "~/types/conversation.types";
 
@@ -30,27 +33,6 @@ export default function ConversationSidebar({
   // Ensure conversations is always an array
   const safeConversations = conversations || [];
 
-  const formatDate = (date: Date | string) => {
-    const dateObj = typeof date === "string" ? new Date(date) : date;
-    const now = new Date();
-    const diffInHours = (now.getTime() - dateObj.getTime()) / (1000 * 60 * 60);
-
-    if (diffInHours < 1) {
-      return "Just now";
-    } else if (diffInHours < 24) {
-      return `${Math.floor(diffInHours)}h ago`;
-    } else if (diffInHours < 168) {
-      return `${Math.floor(diffInHours / 24)}d ago`;
-    } else {
-      return dateObj.toLocaleDateString();
-    }
-  };
-
-  const truncateTitle = (title: string, maxLength: number = 30) => {
-    return title.length > maxLength
-      ? title.substring(0, maxLength) + "..."
-      : title;
-  };
 
   const handleDeleteClick = (e: React.MouseEvent, conversationId: string) => {
     e.preventDefault();
@@ -215,7 +197,7 @@ export default function ConversationSidebar({
                     <div className="space-y-1">
                       <div className="flex items-start justify-between">
                         <h3 className="font-medium text-gray-900 dark:text-gray-100 text-sm leading-tight flex-1 pr-2">
-                          {truncateTitle(conversation.title)}
+                          {truncateText(conversation.title)}
                         </h3>
                         <div className="flex items-center space-x-1">
                           <span className="text-xs text-gray-500 dark:text-gray-400">
@@ -226,7 +208,7 @@ export default function ConversationSidebar({
                               handleDeleteClick(e, conversation.id)
                             }
                             className="opacity-0 group-hover:opacity-100 p-1 rounded hover:bg-red-100 dark:hover:bg-red-900/20 transition-all duration-200"
-                            aria-label={`Delete conversation: ${truncateTitle(conversation.title, 20)}`}
+                            aria-label={`Delete conversation: ${truncateText(conversation.title, 20)}`}
                           >
                             {deletingConversationId === conversation.id ? (
                               <svg
@@ -261,7 +243,7 @@ export default function ConversationSidebar({
                         </div>
                       </div>
                       <p className="text-xs text-gray-500 dark:text-gray-400">
-                        {formatDate(conversation.updatedAt)}
+                        {formatRelativeDate(conversation.updatedAt)}
                       </p>
                     </div>
                   )}
@@ -271,11 +253,11 @@ export default function ConversationSidebar({
                 {showDeleteConfirm === conversation.id && !isCollapsed && (
                   <div
                     role="dialog"
-                    aria-label={`Delete conversation: ${truncateTitle(conversation.title, 20)}`}
+                    aria-label={`Delete conversation: ${truncateText(conversation.title, 20)}`}
                     className="absolute top-0 left-0 right-0 bg-white dark:bg-gray-800 border border-red-200 dark:border-red-800 rounded-lg p-3 shadow-lg z-10"
                   >
                     <p className="text-sm text-gray-900 dark:text-gray-100 mb-3">
-                      Delete &ldquo;{truncateTitle(conversation.title, 20)}
+                      Delete &ldquo;{truncateText(conversation.title, 20)}
                       &rdquo;?
                     </p>
                     <div className="flex space-x-2">
